Tweaks and simplifications
